Prepare mupen64plus.cfg before starting emulation
* disable hotkeys unsuitable for online play
* adjust screenshot and savegame paths to point to correct directory
* split Mupen64Plus code into multiple files
