+++ src/core/emulator/m64plus.cpp
@@ -15,12 +15,6 @@
 namespace Core::Emulator
 {
 
-template<typename... TArgs>
-bool all_true(const TArgs& ... args)
-{
-    return (args && ...);
-}
-
 namespace
 {
 
@@ -30,344 +24,6 @@ bool failed(Mupen64Plus::Error err)
 }
 
 }
-
-namespace M64PlusHelper
-{
-
-Core::Core(std::string config_path, std::string data_path)
-:Core(get_current_process(), std::move(config_path), std::move(data_path))
-{
-    init_symbols();
-    init_core();
-}
-
-Core::Core(dynlib_t lib, std::string config_path, std::string data_path)
-:handle_{lib}, config_path_{std::move(config_path)}, data_path_{std::move(data_path)}
-{
-    init_symbols();
-    init_core();
-}
-
-Core::Core(const std::string& lib_path, std::string config_path, std::string data_path)
-:handle_{load_library(lib_path.c_str())}, config_path_{std::move(config_path)}, data_path_{std::move(data_path)}
-{
-    if(!handle_.lib)
-    {
-        // Library file does not exist
-        logger()->error("Failed to open library file: \"{}\"", get_lib_error_msg());
-        throw std::system_error(make_error_code(Error::LIB_LOAD_FAILED));
-    }
-    init_symbols();
-    init_core();
-}
-
-Core::Core(Core&& other) noexcept
-:handle_{std::move(other.handle_)}, fn_{other.fn_},
-config_path_{other.config_path_}, data_path_{other.data_path_}
-{
-}
-
-Core& Core::operator=(Core&& other) noexcept
-{
-    swap(*this, other);
-
-    return *this;
-}
-
-Core::~Core()
-{
-    if(!handle_.lib)
-        return;
-
-    destroy_core();
-}
-
-void swap(Core& first, Core& second) noexcept
-{
-    using std::swap;
-
-    swap(first.handle_, second.handle_);
-    swap(first.fn_, second.fn_);
-    swap(first.config_path_, second.config_path_);
-    swap(first.data_path_, second.data_path_);
-}
-
-void Core::attach_plugin(Plugin& plugin)
-{
-    auto ret{fn_.core_attach_plugin(plugin.info().type, plugin.handle())};
-    if(failed(ret))
-    {
-        // Failed to attach plugin
-        auto errc{make_error_code(ret)};
-        logger()->error("Failed to attach plugin {}: {}", plugin.info().name, errc.message());
-        throw std::system_error(errc);
-    }
-}
-
-void Core::detach_plugin(M64PTypes::m64p_plugin_type type)
-{
-    auto ret{fn_.core_detach_plugin(type)};
-    if(failed(ret))
-    {
-        // Failed to correctly detach plugin
-        auto errc{make_error_code(ret)};
-        logger()->warn("Failed to correctly detach plugin of type {}: {}", type, errc.message());
-    }
-}
-
-void* Core::get_mem_ptr()
-{
-    return fn_.debug_get_mem_ptr(M64PTypes::M64P_DBG_PTR_RDRAM);
-}
-
-Error Core::do_cmd(M64PTypes::m64p_command cmd, int p1, void* p2)
-{
-    return fn_.core_do_cmd(cmd, p1, p2);
-}
-
-void Core::init_symbols()
-{
-    resolve_symbol(fn_.plugin_get_version, "PluginGetVersion");
-    resolve_symbol(fn_.core_startup, "CoreStartup");
-    resolve_symbol(fn_.core_shutdown, "CoreShutdown");
-    resolve_symbol(fn_.core_attach_plugin, "CoreAttachPlugin");
-    resolve_symbol(fn_.core_detach_plugin, "CoreDetachPlugin");
-    resolve_symbol(fn_.core_do_cmd, "CoreDoCommand");
-    resolve_symbol(fn_.debug_get_mem_ptr, "DebugMemGetPointer");
-
-    if(!all_true(fn_.plugin_get_version, fn_.core_attach_plugin, fn_.core_detach_plugin, fn_.core_startup,
-                 fn_.core_shutdown, fn_.core_do_cmd, fn_.debug_get_mem_ptr))
-    {
-        auto errc{make_error_code(Error::SYM_NOT_FOUND)};
-        logger()->error("Failed to resolve symbols of core");
-        throw std::system_error(errc);
-    }
-}
-
-void Core::init_core()
-{
-    const char* name_ptr{};
-    auto ret{
-    fn_.plugin_get_version(&info_.type, &info_.plugin_version, &info_.api_version, &name_ptr, &info_.capabilities)};
-    if(failed(ret))
-    {
-        // Failed to get core info
-        auto errc{make_error_code(ret)};
-        logger()->error("Failed to retrieve core info: {}", errc.message());
-        throw std::system_error(errc);
-    }
-
-    info_.name = name_ptr;
-    ret = fn_.core_startup(CORE_API_VERSION, config_path_.c_str(), data_path_.c_str(), nullptr,
-                           nullptr, nullptr, nullptr);
-    if(failed(ret))
-    {
-        // Failed to startup core
-        info_ = {};
-        auto errc{make_error_code(ret)};
-        logger()->error("Failed to start {} v{}, api: {} (capabilities: {:#x}): {}", name_ptr, info_.plugin_version,
-                        info_.api_version, info_.capabilities, errc.message());
-        throw std::system_error(errc);
-    }
-
-    logger()->info("Initialized {} v{}, api: {:#x} (capabilities: {:#x})", name_ptr, info_.plugin_version,
-                   info_.api_version, info_.capabilities);
-}
-
-void Core::destroy_core()
-{
-    auto ret{fn_.core_shutdown()};
-    if(failed(ret))
-    {
-        // Failed to correctly shutdown core
-        auto errc{make_error_code(ret)};
-        logger()->warn("Failed to correctly shutdown core: ", errc.message());
-    }
-    logger()->info("Shutdown {}", Plugin::type_str(M64PTypes::M64PLUGIN_CORE));
-}
-
-dynlib_t Core::handle()
-{
-    return handle_.lib;
-}
-
-const PluginInfo& Core::info() const
-{
-    return info_;
-}
-
-
-Plugin::Plugin(Core& core, dynlib_t lib)
-:handle_{lib}
-{
-    init_symbols();
-    init_plugin(core.handle());
-}
-
-Plugin::Plugin(Core& core, const std::string& lib_path)
-:handle_{load_library(lib_path.c_str())}
-{
-    if(!handle_.lib)
-    {
-        // Library file does not exist
-        logger()->error("Failed to open library file: \"{}\"", get_lib_error_msg());
-        throw std::system_error(make_error_code(Error::LIB_LOAD_FAILED));
-    }
-    init_symbols();
-    init_plugin(core.handle());
-}
-
-Plugin::Plugin(Plugin&& other) noexcept
-:handle_{std::move(other.handle_)}, fn_{other.fn_}, info_{other.info_}
-{
-}
-
-Plugin& Plugin::operator=(Plugin&& other) noexcept
-{
-    swap(*this, other);
-
-    return *this;
-}
-
-Plugin::~Plugin()
-{
-    if(!handle_.lib)
-        return;
-
-    destroy_plugin();
-}
-
-void swap(Plugin& first, Plugin& second) noexcept
-{
-    using std::swap;
-
-    swap(first.handle_, second.handle_);
-    swap(first.info_, second.info_);
-    swap(first.fn_, second.fn_);
-}
-
-const PluginInfo& Plugin::info() const
-{
-    return info_;
-}
-
-dynlib_t Plugin::handle()
-{
-    return handle_.lib;
-}
-
-PluginInfo Plugin::get_plugin_info(const std::string& file)
-{
-    auto lib{load_library(file.c_str())};
-
-	if(!lib)
-	{
-		logger()->debug("Failed to load file {} as library: {}", file, get_lib_error_msg());
-		return {};
-	}
-
-    auto ret{get_plugin_info(lib)};
-    free_library(lib);
-    return ret;
-}
-
-PluginInfo Plugin::get_plugin_info(dynlib_t lib)
-{
-    PluginInfo info{};
-    auto get_version{reinterpret_cast<plugin_get_version_t>(get_symbol(lib, "PluginGetVersion"))};
-    if(!get_version)
-        return {};
-
-    const char* name_ptr{};
-    get_version(&info.type, &info.plugin_version, &info.api_version, &name_ptr, &info.capabilities);
-    info.name = name_ptr;
-
-    return info;
-}
-
-const char* Plugin::type_str(M64PTypes::m64p_plugin_type type_id)
-{
-    using namespace M64PTypes;
-
-    switch(type_id)
-    {
-    case M64PLUGIN_RSP:
-        return "RSP";
-    case M64PLUGIN_GFX:
-        return "VIDEO";
-    case M64PLUGIN_AUDIO:
-        return "AUDIO";
-    case M64PLUGIN_INPUT:
-        return "INPUT";
-    case M64PLUGIN_CORE:
-        return "CORE";
-    case M64PLUGIN_NULL:
-        return "NULL";
-    default:
-        return "Invalid plugin type";
-    }
-}
-
-void Plugin::init_symbols()
-{
-    resolve_symbol(fn_.startup, "PluginStartup");
-    resolve_symbol(fn_.shutdown, "PluginShutdown");
-    resolve_symbol(fn_.get_version, "PluginGetVersion");
-
-    if(!all_true(fn_.startup, fn_.shutdown, fn_.get_version))
-    {
-        // Failed to resolve symbol
-        auto errc{make_error_code(Error::SYM_NOT_FOUND)};
-        logger()->error("Failed to resolve symbols of plugin");
-        throw std::system_error(errc);
-    }
-}
-
-void Plugin::init_plugin(dynlib_t core_lib)
-{
-    const char* name_ptr{};
-    auto ret{fn_.get_version(&info_.type, &info_.plugin_version, &info_.api_version, &name_ptr, &info_.capabilities)};
-    if(failed(ret))
-    {
-        // Failed to retrieve version info
-        auto errc{make_error_code(ret)};
-        logger()->error("Failed to retrieve core info: {}", errc.message());
-        info_ = {};
-        throw std::system_error(errc);
-    }
-
-    ret = fn_.startup(core_lib, nullptr, nullptr);
-    if(failed(ret))
-    {
-        // Plugin failed to initialize
-        auto errc{make_error_code(ret)};
-        logger()->error("Failed to start {} plugin {} v{}, api: {} (capabilities: {:#x}): {}",
-                        Plugin::type_str(info_.type), name_ptr, info_.plugin_version, info_.api_version,
-                        info_.capabilities, errc.message());
-        info_ = {};
-        throw std::system_error(errc);
-    }
-
-    info_.name = name_ptr;
-
-    logger()->info("Initialized {} plugin {} v{}, api: {:#x} (capabilities: {:#x})",
-                   Plugin::type_str(info_.type), name_ptr, info_.plugin_version, info_.api_version, info_.capabilities);
-}
-
-void Plugin::destroy_plugin()
-{
-    auto ret{fn_.shutdown()};
-    if(failed(ret))
-    {
-        // Failed to destroy plugin
-        auto errc{make_error_code(ret)};
-        logger()->warn("Failed to correctly shutdown plugin {}: {}", info_.name, errc.message());
-    }
-
-    logger()->info("Shutdown {} plugin", Plugin::type_str(info_.type));
-}
-
-} // M64PlusHelper
 
 Mupen64Plus::Mupen64Plus(Core&& core)
 :core_{std::move(core)}
@@ -449,6 +105,7 @@ void Mupen64Plus::unload_rom()
 void Mupen64Plus::execute()
 {
     running_ = true;
+    core_.prepare_config_file();
     attach_plugins();
     logger()->info("Started n64 emulation");
     auto ret{core_.do_cmd(M64PTypes::M64CMD_EXECUTE, 0, nullptr)};
