+++ src/qt_gui/main.cpp
@@ -30,8 +30,7 @@ static void create_app_dirs(const AppSettings& settings)
     try
     {
         fs::create_directories(settings.main_config_dir());
-        fs::create_directories(settings.m64p_data_dir());
-        fs::create_directories(settings.m64p_config_dir());
+        fs::create_directories(settings.m64p_dir());
     }
     catch(const std::exception& e)
     {
